refactoring modal to its own component
